upd: labPBR support / Material presets (#3)
* refactor: Split into multiple files

- Added registry, setups, and teardowns to control plugin actions

* feat: Reset exposure Action

- Added messages to baking failures
- Added reset exposure button
- Added changelog build script

* feat: labPBR output Action

* feat: Decode labPBR channels

* chore: Remove unused code

- Revert normal map usage in PbrMaterial.getMaterial()

* fix: labPBR specular map logic

- Fixed bug in alpha channel
- Added better specular and normal export names

* feat: Material brush presets

* feat: Channel assignment panel

- Created panel to list current channel assignments

* upd: Channel component cleanup

* upd: Style channel list

* feat: Material brush presets library

- Created Dialog for editing, saving and loading material brush presets

* upd: Change Action icon

- Changed material brush Action icon

* upd: Channel assignment condition
