+++ src/pbr_preview/src/deps.ts
@@ -1,0 +1,7 @@
+// @ts-expect-error THREE UMD is OK
+const three = THREE;
+// @ts-expect-error Vue UMD is OK
+const vue = Vue;
+const bb = Blockbench;
+
+export { three, bb, vue };
