+++ src/pbr_preview/src/lib/panels/index.ts
@@ -1,0 +1,3 @@
+import "./materialBrush";
+import "./pbrSettings";
+import "./channels";
