+++ src/pbr_preview/src/lib/tools/index.ts
@@ -1,0 +1,1 @@
+import "./materialBrush";
